Simplify user session logic + redirect user when using OAuth code
